+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 Version 0.2.0
 -------------
 
+*In development*
 
 - Values in :class:`~relief.Form` structures can now be validated by
   subclassing :class:`~relief.Form` and adding `validate_{key}` methods.
@@ -12,6 +13,9 @@ Version 0.1.1
 -------------
 
 *In development*
+
+- Removed mention of being able to mutate :class:`relief.Dict` and
+  :class:`relief.Form`.
 
 Version 0.1.0
 -------------